Refactor: Replace System.IO usage with Utils.System.IO
- Removed direct usage of System.IO and System.Exit in multiple files.
- Introduced Utils.System.IO to handle IO operations and exit codes.
- Updated all relevant imports and function calls to use the new Utils.System.IO module.
- Deleted obsolete System.Exit and System.IO modules.
- Cleaned up unused encoders and decoders related to HTTP and exceptions.
- Consolidated error handling by replacing exit calls with crash utility.

+++ run-make.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+clear_cache=true;
+
 set -o errexit;
 set -o nounset;
 
@@ -12,24 +14,30 @@ set -o nounset;
 # guida-lang/stdlib source code:
 
 export GUIDA_HOME="$PWD/.guida";
-rm -rf "$GUIDA_HOME" && mkdir -p "$GUIDA_HOME";
+if [ "$clear_cache" = true ]; then
+    rm -rf "$GUIDA_HOME" && mkdir -p "$GUIDA_HOME";
+fi
+
 rm -rf guida-stuff;
 
 # Create a link to the git package
-CORE_LINK="${GUIDA_HOME}/1.0.0/packages/guida-lang/stdlib/2.0.1"
+CORE_LINK="${GUIDA_HOME}/1.0.0/packages/guida-lang/stdlib/1.0.1"
 CORE_GIT_DIR="/Users/decio/code/guida-lang/stdlib"
 
-echo;
-echo "Linking $CORE_LINK to $CORE_GIT_DIR"
-echo;
-mkdir -p "$(dirname $CORE_LINK)"
-ln -sv "${CORE_GIT_DIR}" "${CORE_LINK}"
-rm -vf "${CORE_GIT_DIR}"/*.dat "${CORE_GIT_DIR}"/doc*.json
-
+if [ "$clear_cache" = true ]; then
+    echo;
+    echo "Linking $CORE_LINK to $CORE_GIT_DIR"
+    echo;
+    mkdir -p "$(dirname $CORE_LINK)"
+    ln -sv "${CORE_GIT_DIR}" "${CORE_LINK}"
+    rm -vf "${CORE_GIT_DIR}"/*.dat "${CORE_GIT_DIR}"/doc*.json
+fi
 # Now we can run the make against the symlinked source code for real
 
 echo;
 echo "running make ...";
 echo;
 
-time GUIDA_REGISTRY=http://localhost:3000 ../compiler-master/bin/index.js make src/Terminal/Main.elm;
+time ../compiler-master/bin/index.js make --output=bin/guida.js src/Terminal/Main.elm;
+# time bin/index.js make --output=bin/guida.js src/Terminal/Main.elm;
+node scripts/replacements.js bin/guida.js
